+++ wazo_plugind/helpers/tests/test_validator.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-3.0+
 
 from unittest import TestCase
-from hamcrest import assert_that, calling, has_properties
+from hamcrest import assert_that, calling, has_properties, not_
 from mock import ANY, Mock, patch
 from xivo_test_helpers.hamcrest.raises import raises
 
@@ -17,9 +17,11 @@ CURRENT_WAZO_VERSION = '17.10'
 class TestValidator(TestCase):
 
     def setUp(self):
-        self.validator = Validator(Mock(), CURRENT_WAZO_VERSION)
+        self.plugin_db = Mock()
 
     def test_that_missing_fields_raises(self):
+        validator = self.new_validator()
+
         metadata = self.new_metadata()
         metadata.pop('version', None)
 
@@ -29,7 +31,7 @@ class TestValidator(TestCase):
                         'message': ANY}
         }
         assert_that(
-            calling(self.validator.validate).with_args(metadata),
+            calling(validator.validate).with_args(metadata),
             raises(exceptions.PluginValidationException).matching(
                 has_properties('error_id', 'validation_error',
                                'message', 'Validation error',
@@ -38,6 +40,8 @@ class TestValidator(TestCase):
         )
 
     def test_that_invalid_name_raises(self):
+        validator = self.new_validator()
+
         metadata = self.new_metadata(name='no_underscore_allowed')
 
         expected_details = {
@@ -46,7 +50,7 @@ class TestValidator(TestCase):
                      'message': ANY}
         }
         assert_that(
-            calling(self.validator.validate).with_args(metadata),
+            calling(validator.validate).with_args(metadata),
             raises(exceptions.PluginValidationException).matching(
                 has_properties('error_id', 'validation_error',
                                'message', 'Validation error',
@@ -55,6 +59,8 @@ class TestValidator(TestCase):
         )
 
     def test_that_invalid_namespace_raises(self):
+        validator = self.new_validator()
+
         metadata = self.new_metadata(namespace='no-dash-allowed')
 
         expected_details = {
@@ -63,7 +69,7 @@ class TestValidator(TestCase):
                           'message': ANY}
         }
         assert_that(
-            calling(self.validator.validate).with_args(metadata),
+            calling(validator.validate).with_args(metadata),
             raises(exceptions.PluginValidationException).matching(
                 has_properties('error_id', 'validation_error',
                                'message', 'Validation error',
@@ -72,6 +78,8 @@ class TestValidator(TestCase):
         )
 
     def test_plugin_format_version_from_the_future(self):
+        validator = self.new_validator()
+
         metadata = self.new_metadata(plugin_format_version=_MAX_PLUGIN_FORMAT_VERSION + 1)
 
         expected_details = {
@@ -79,7 +87,7 @@ class TestValidator(TestCase):
                                       'constraint': {'min': 0, 'max': _MAX_PLUGIN_FORMAT_VERSION},
                                       'message': ANY}}
         assert_that(
-            calling(self.validator.validate).with_args(metadata),
+            calling(validator.validate).with_args(metadata),
             raises(exceptions.PluginValidationException).matching(
                 has_properties('error_id', 'validation_error',
                                'message', 'Validation error',
@@ -88,6 +96,8 @@ class TestValidator(TestCase):
         )
 
     def test_max_wazo_version_too_small(self):
+        validator = self.new_validator()
+
         metadata = self.new_metadata(max_wazo_version='16.16')
 
         expected_details = {
@@ -95,7 +105,7 @@ class TestValidator(TestCase):
                                  'constraint': {'min': CURRENT_WAZO_VERSION},
                                  'message': ANY}}
         assert_that(
-            calling(self.validator.validate).with_args(metadata),
+            calling(validator.validate).with_args(metadata),
             raises(exceptions.PluginValidationException).matching(
                 has_properties('error_id', 'validation_error',
                                'message', 'Validation error',
@@ -104,6 +114,8 @@ class TestValidator(TestCase):
         )
 
     def test_min_wazo_version_too_high(self):
+        validator = self.new_validator()
+
         metadata = self.new_metadata(min_wazo_version='17.11')
 
         expected_details = {
@@ -111,7 +123,7 @@ class TestValidator(TestCase):
                                  'constraint': {'max': CURRENT_WAZO_VERSION},
                                  'message': ANY}}
         assert_that(
-            calling(self.validator.validate).with_args(metadata),
+            calling(validator.validate).with_args(metadata),
             raises(exceptions.PluginValidationException).matching(
                 has_properties('error_id', 'validation_error',
                                'message', 'Validation error',
@@ -120,11 +132,18 @@ class TestValidator(TestCase):
         )
 
     def test_plugin_already_installed(self):
+        validator = self.new_validator()
         metadata = self.new_metadata()
-
-        with patch.object(self.validator._db, 'is_installed', return_value=True):
-            assert_that(calling(self.validator.validate).with_args(metadata),
+        with patch.object(validator._db, 'is_installed', return_value=True):
+            assert_that(calling(validator.validate).with_args(metadata),
                         raises(exceptions.PluginAlreadyInstalled))
+
+        validator = self.new_validator(install_params={'reinstall': True})
+        metadata = self.new_metadata()
+        with patch.object(validator._db, 'is_installed', return_value=True):
+            assert_that(
+                calling(validator.validate).with_args(metadata),
+                not_(raises(exceptions.PluginAlreadyInstalled)))
 
     def new_metadata(self, name='valid-name', namespace='validns', version='0.0.1', **kwargs):
         metadata = dict(kwargs)
@@ -132,3 +151,8 @@ class TestValidator(TestCase):
         metadata['namespace'] = namespace
         metadata['version'] = version
         return metadata
+
+    def new_validator(self, plugin_db=None, wazo_version=CURRENT_WAZO_VERSION, install_params=None):
+        plugin_db = plugin_db or self.plugin_db
+        install_params = install_params or {'reinstall': False}
+        return Validator(plugin_db, wazo_version, install_params)