+++ wazo_plugind/http.py
@@ -16,10 +16,12 @@ from xivo.rest_api_helpers import handle_api_exception
 from .schema import (
     MarketListRequestSchema,
     MarketListResultSchema,
+    PluginInstallQueryStringSchema,
     PluginInstallSchema,
 )
 from .exceptions import (
     InvalidInstallParamException,
+    InvalidInstallQueryStringException,
     InvalidListParamException,
     MarketNotFoundException,
 )
@@ -126,13 +128,13 @@ class Plugins(_AuthentificatedResource):
         if errors:
             raise InvalidInstallParamException(errors)
 
-        method, options = body['method'], body['options']
-        return self._post(method, options)
+        params, errors = PluginInstallQueryStringSchema().load(request.args)
+        if errors:
+            raise InvalidInstallQueryStringException(errors)
 
-    def _post(self, method, options):
-        uuid = self.plugin_service.create(method, **options)
+        uuid = self.plugin_service.create(body['method'], params, body['options'])
 
-        return {'uuid': uuid}
+        return dict(uuid=uuid)
 
     @classmethod
     def add_resource(cls, api, *args, **kwargs):